refactor(eventbus): split process_manager.rs into process_manager/ submodules
